fix: update callback package initializer
